+++ lib/src/inner_logger_operators_mixin.dart
@@ -1,0 +1,24 @@
+import 'inner_logger.dart';
+
+/// {@nodoc}
+mixin InnerLoggerOperatorsMixin on InnerLogger {
+  @override
+  void operator +(int v) {
+    // TODO: implement +
+  }
+
+  @override
+  void operator -(int v) {
+    // TODO: implement -
+  }
+
+  @override
+  void operator <(Object info) {
+    // TODO: implement <
+  }
+
+  @override
+  void operator <<(Object debug) {
+    // TODO: implement <<
+  }
+}